requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astro-sdk-python
airflow-provider-duckdb
apache-airflow-providers-amazon
dbt-core
dbt-duckdb
astronomer-cosmos

awscli==1.32.84
boto3==1.34.84
botocore==1.34.84
colorama==0.4.4
docutils==0.16
jmespath==1.0.1
pyasn1==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1